% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_grid_pts.R
\name{create_grid_pts}
\alias{create_grid_pts}
\title{Create grid of points that fall within the contiguous USA.}
\usage{
create_grid_pts(resolution = 1)
}
\arguments{
\item{resolution}{number: increment of the sequence of longitude and
latitude grid points}
}
\value{
a data-frame with the longitudes and latitudes of grid points within
contiguous USA based on specified resolution.

a Dataframe containing the following columns, where a row
represents a grid point:
\itemize{
  \item \code{LONGITUDE} longitude of grid points
  \item \code{LATITUDE} latitude of grid points
  \item \code{inside_usa} boolean vector for whether the grid point is
  inside the contiguous USA
}
}
\description{
Return a Dataframe containing longitude and latitudes
of grid points within the contiguous USA based on the resolution specified.
}
\examples{
# create grid points with high resolution
grid_pts <- create_grid_pts(1)
plot(grid_pts[,c(1,2)][grid_pts$inside_usa,])
}
