% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_station_data.R
\name{interpolate_station_data}
\alias{interpolate_station_data}
\title{Interpolate data from the stations to grid points within the contiguous USA}
\usage{
interpolate_station_data()
}
\value{
a Dataframe containing the following columns
\itemize{
  \item \code{WBANNO} station WBAN number
  \item \code{station_name} name of the USCRN station
  \item \code{LONGITUDE} Station longitude
  \item \code{LATITUDE} Station latitude
}
}
\description{
This function interpolates all the stations to grid points by returning the
stations that are within the contiguous USA.
}
\examples{
# Get yearly cycle for station NC_Asheville_8_SSW
interpolations <- interpolate_station_data()
plot(interpolations$LONGITUDE, interpolations$LATITUDE)
plot_interpolations(interpolations, create_grid_pts())
}
