% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearly_cycle.R
\name{yearly_cycle}
\alias{yearly_cycle}
\title{Estimate the yearly cycle for one station}
\usage{
yearly_cycle(station_daily_data)
}
\arguments{
\item{station_daily_data}{daily weather data for a specific USCRN station id}
}
\value{
a Dataframe containing the following columns:
\itemize{
  \item day: day number (1-366)
  \item expected_avg_temp: expected average temperature on each day
}
}
\description{
This function estimates a station's yearly cycle, which is simply the
expected temperature on each day of the year. The estimation occurs by using
sine and cosine terms in the model to predict expected temperature. Returns
a dataframe containing expected average temperature of each day of the year.
}
\examples{
# Get yearly cycle for station NC_Asheville_8_SSW
station_id <- 53877
station_daily_data <- daily_weather_data[daily_weather_data$WBANNO == station_id, ]
yearly_cycle(station_daily_data)
}
