% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_estimate.R
\name{trend_estimate}
\alias{trend_estimate}
\title{Estimate the trend of temperatures over time}
\usage{
trend_estimate()
}
\description{
This function calculates the trend of temperatures over time, in units of degrees Celsius per year. Returns a dataframe
containing expected average temperature of each day of the year.
}
